backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.8.0
sqlalchemy>=2.0.25
PyJWT>=2.8.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
psycopg2-binary
slowapi>=0.1.9